requirements.txt
biopython==1.80
biotite==0.37.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
ete3==3.1.3
fonttools==4.40.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
memory-profiler==0.61.0
msgpack==1.0.5
networkx==3.1
numpy==1.24.1
packaging==23.1
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tqdm==4.65.0
urllib3==2.0.3
zipp==3.15.0
